requirements.txt
fastapi==0.116.1
uvicorn==0.27.1
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
vosk==0.3.45
numpy==1.26.4
soundfile==0.12.1
openai==1.3.0
pypdf==4.2.0
chromadb==0.5.5
tiktoken==0.7.0

